# expression: implement vectorized evaluation for `builtinGreatestTimeSig`