PCP #12101
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinGreatestTimeSig, for #12106
### What is changed and how it works?
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinCompareFunc -run=BenchmarkVectorizedBuiltinCompareFunc -args 'builtinGreatestTimeSig'
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinCompareFunc/builtinGreatestTimeSig-VecBuiltinFunc-4         	     706	   1605207 ns/op	  552693 B/op	   31285 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGreatestTimeSig-NonVecBuiltinFunc-4      	     700	   1685922 ns/op	  553288 B/op	   30877 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGreatestTimeSig-VecBuiltinFunc#01-4      	     452	   2669034 ns/op	  927032 B/op	   51971 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGreatestTimeSig-NonVecBuiltinFunc#01-4   	     548	   2108409 ns/op	  704447 B/op	   39027 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	5.538s
```
### Check List <!--REMOVE the items that are not applicable-->
Tests <!-- At least one of them must be included. -->
- Unit test