@mmyj 2019-10-21
benchmark of new implement
```
go test -v -benchmem -bench=BenchmarkVectorizedBuiltinCompareFunc -run=BenchmarkVectorizedBuiltinCompareFunc -args 'builtinGreatestTimeSig'
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinCompareFunc/builtinGreatestTimeSig-VecBuiltinFunc-4         	     537	   2146397 ns/op	  737807 B/op	   41654 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGreatestTimeSig-NonVecBuiltinFunc-4      	     597	   1987401 ns/op	  659135 B/op	   36580 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	2.789s
```